Move admin_url_namespace and base_url_path to ModelViewSet class
